notebooks: allow to test GeoServer on new production server
Switch from 'boreas.ouranos.ca' to 'pavics.ouranos.ca' will allow us to
conditionally regex replace 'pavics.ouranos.ca' to the server under test
so we can actually test the Geoserver on that new server.

Everywhere under this pavics-sdi repo, we use 'pavics.ouranos.ca'.  This
is the only locations that hardcode 'boreas.ouranos.ca' because
geoserver data are not replicated to standard test servers so hardcoding
it this way will use the data from the production server.

This patch achieve the same "data from prod" but also allow override if
we need to actually test the future production server.